src/domain/notification/application/use-cases/read-notification.ts
import { Either, left, right } from '@/core/either'
import { NotAllowedError } from '@/core/errors/not-allowed-error'
import { ResourceNotFoundError } from '@/core/errors/resource-not-found-error'
import { Notification } from '../../enterprise/entities/notification'
import { NotificationsRepository } from '../repositories/notifications-repository'

interface ReadNotificationUseCaseInput {
  notificationId: string
  recipientId: string
}

type ReadNotificationUseCaseOutput = Promise<
  Either<
    ResourceNotFoundError | NotAllowedError,
    {
      notification: Notification
    }
  >
>

export class ReadNotificationUseCase {
  constructor(private notificationsRepository: NotificationsRepository) {}

  async execute({ notificationId, recipientId }: ReadNotificationUseCaseInput): ReadNotificationUseCaseOutput {
    const notification = await this.notificationsRepository.findById(notificationId)

    if (!notification) {
      return left(new ResourceNotFoundError())
    }

    if (recipientId !== notification.recipientId.value) {
      return left(new NotAllowedError())
    }

    notification.read()

    await this.notificationsRepository.save(notification)

    return right({ notification })
  }
}
